Fixed script for version check

+++ src/getAppVersion.sh
@@ -23,5 +23,8 @@ $MALINE/lib/apktool/apktool decode $1 -o $TMP_DIR 1>/dev/null 2>/dev/null
 filename=$TMP_DIR/apktool.yml
 minSdkVersion=$(grep -e "minSdkVersion" $filename | awk -F ": " '{ print $2 }' | awk -F "'" '{ print $2 }')
 targetSdkVersion=$(grep -e "targetSdkVersion" $filename | awk -F ": " '{ print $2 }' | awk -F "'" '{ print $2 }')
-rm -rf $TMP_DIR
+status=$?
+if [ $status -eq 0 ]; then
+    rm -rf $TMP_DIR
+fi
 echo $minSdkVersion,$targetSdkVersion
